feat: Update package dependencies and add daisyui
The code changes update the package dependencies in the package-lock.json and package.json files. Specifically, the "daisyui" package is added as a dependency with version 4.12.2. This addition enables the usage of the DaisyUI library, which provides a set of utility classes for Tailwind CSS. This update enhances the project by leveraging the features and improvements offered by the latest version of the "daisyui" package.