feat: Add dropdown menu to navigation bar
The code changes include adding a dropdown menu to the navigation bar on the homepage. This dropdown menu provides a list of items that can be accessed by clicking on the hamburger menu icon. This enhances the user experience by organizing and presenting additional navigation options in a compact and intuitive manner.